Move filterer now that tagged is out
