+++ crates/cli/src/filterer/dirs.rs
